// timeline-viewer/types.ts
export interface TEvent {
  id: string;
  name: string;
  date: string; // Using string for simplicity, can be parsed to Date
  description: string;
  pictureUrl: string;
  people: string[];
  tags: string[];
}

export interface TTimeline {
  id: string;
  name:string;
  description: string;
  events: TEvent[];
}